+++ src/codeauditor.ts
@@ -2,8 +2,9 @@ import * as vscode from 'vscode';
 import { triggerUpdateDecorations } from './decorators';
 import { toggleFilter } from './filter';
 import { noteProvider } from './noteexplorer';
-import { noteState, noteType } from './types';
-import { newNote, removeNote, setNoteState, setNoteType } from './notes';
+import { fileState, noteState, noteType } from './types';
+import { excludePath, newNote, removeNote, setFileState, setNoteState, setNoteType } from './notes';
+import { progressProvider } from './progressexplorer';
 import { generateReport } from './report';
 import { auditDataInit } from './storage';
 import { ImportSemgrepReport, ImportSlitherReport } from './importnotes';
@@ -16,41 +17,50 @@ export function activate(context: vscode.ExtensionContext) {
 	const noteExplorer : noteProvider = new noteProvider();
 	vscode.window.registerTreeDataProvider('code-auditor.noteExplorer', noteExplorer);
 
+	const progressExplorer : progressProvider = new progressProvider();
+	vscode.window.registerTreeDataProvider('code-auditor.progressExplorer', progressExplorer);
+
 	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.newNote', (line : string) => {
 		newNote(line);
 	}));
 	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.removeNote', (line : string) => {
 		removeNote(line);
 	}));
-	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.notesOn', (line : string) => {
+	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.notesOn', () => {
 		toggleFilter(noteType.Note);
 	}));
-	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.notesOff', (line : string) => {
+	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.notesOff', () => {
 		toggleFilter(noteType.Note);
 	}));
-	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.issueOn', (line : string) => {
+	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.issueOn', () => {
 		toggleFilter(noteType.Issue);
 	}));
-	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.issueOff', (line : string) => {
+	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.issueOff', () => {
 		toggleFilter(noteType.Issue);
 	}));
-	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.openOn', (line : string) => {
+	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.openOn', () => {
 		toggleFilter(noteState.Open);
 	}));
-	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.openOff', (line : string) => {
+	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.openOff', () => {
 		toggleFilter(noteState.Open);
 	}));
-	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.confirmedOn', (line : string) => {
+	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.confirmedOn', () => {
 		toggleFilter(noteState.Confirmed);
 	}));
-	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.confirmedOff', (line : string) => {
+	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.confirmedOff', () => {
 		toggleFilter(noteState.Confirmed);
 	}));
-	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.discardedOn', (line : string) => {
+	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.discardedOn', () => {
 		toggleFilter(noteState.Discarded);
 	}));
-	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.discardedOff', (line : string) => {
+	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.discardedOff', () => {
 		toggleFilter(noteState.Discarded);
+	}));
+	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.progressOn', () => {
+		toggleFilter(fileState.Reviewed);
+	}));
+	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.filter.progressOff', () => {
+		toggleFilter(fileState.Reviewed);
 	}));
 	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.setNoteState', (args : string) => {
 		if (!args) { return; }
@@ -59,6 +69,15 @@ export function activate(context: vscode.ExtensionContext) {
 	}));
 	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.reopenNote', () => {
 		setNoteState(noteState.Open);
+	}));
+	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.fileStateReview', (item) => {
+		if (item) { setFileState(fileState.Reviewed, item.uri); }
+	}));
+	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.fileStateReOpen', (item) => {
+		if (item) { setFileState(fileState.Pending, item.uri); }
+	}));
+	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.excludePath', (item) => {
+		if (item) { excludePath(item.uri); }
 	}));
 	context.subscriptions.push(vscode.commands.registerCommand('code-auditor.toggleType', () => {
 		setNoteType();